finalKG/requirements.txt
neo4j>=5.0.0
python-dotenv>=1.0.0
pandas>=2.0.0
torch>=2.0.0
torch-geometric>=2.3.0
requests>=2.0.0
tqdm>=4.0.0
